Add generation plugin
